+++ DICOM/Network/Client/States/DicomClientRequestAssociationState.cs
@@ -114,6 +114,7 @@ namespace Dicom.Network.Client.States
             {
                 var dicomRequest = queuedItem.Value;
                 associationToRequest.PresentationContexts.AddFromRequest(dicomRequest);
+                associationToRequest.ExtendedNegotiations.AddFromRequest(dicomRequest);
             }
 
             foreach (var context in _dicomClient.AdditionalPresentationContexts)
@@ -123,6 +124,15 @@ namespace Dicom.Network.Client.States
                     context.UserRole,
                     context.ProviderRole,
                     context.GetTransferSyntaxes().ToArray());
+            }
+            
+            foreach (var extendedNegotiation in _dicomClient.AdditionalExtendedNegotiations)
+            {
+                associationToRequest.ExtendedNegotiations.AddOrUpdate(
+                    extendedNegotiation.SopClassUid,
+                    extendedNegotiation.RequestedApplicationInfo,
+                    extendedNegotiation.ServiceClassUid,
+                    extendedNegotiation.RelatedGeneralSopClasses.ToArray());
             }
 
             await Connection.SendAssociationRequestAsync(associationToRequest).ConfigureAwait(false);